+++ lib/serega/plugins/preloads/validations/check_opt_preload.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class Serega
+  module SeregaPlugins
+    module Preloads
+      class CheckOptPreload
+        class << self
+          def call(opts)
+            return unless opts.key?(:preload)
+
+            raise SeregaError, "Option :preload can not be used together with option :const" if opts.key?(:const)
+          end
+        end
+      end
+    end
+  end
+end